vectordatabase/requirements.txt
torch
open_clip_torch
numpy
Pillow
regex
ftfy
tqdm